Move RawImage to test module for now. Add comments for test blocks.
